test: 서비스 헬스 체크 smoke 테스트 추가
