--- src/test/resources/features/kiosk-service-health.feature
@@ -1,6 +1,0 @@
-Feature: 키오스크 서비스 헬스 체크
-  외부 프로세스로 기동한 키오스크 서비스가 정상적으로 응답하는지 확인한다.
-
-  Scenario: 키오스크 서비스 엔드포인트 확인
-    When 키오스크 서비스 루트 경로에 요청을 보내면
-    Then 200 응답을 받는다
