+++ src/test/resources/features/services-health.feature
@@ -1,0 +1,12 @@
+Feature: 캠핑 서비스 헬스 체크
+  외부 프로세스로 기동한 캠핑 관련 서비스들이 정상적으로 응답하는지 확인한다.
+
+  Scenario Outline: <서비스명> 서비스 엔드포인트 확인
+    When <서비스명> 서비스 루트 경로에 요청을 보낸다
+    Then 200 응답을 받는다
+
+    Examples:
+      | 서비스명      | 포트    |
+      | 키오스크      | 18081  |
+      | 관리자       | 18082  |
+      | 예약        | 18083  |
